move ConditionPredicateAttribute to manager
